Module dex_v3::instruction

Program instructions and their CPI-compatible bindings

Modules

cancel_order

close_market

consume_events

create_market

initialize_account

new_order

settle

sweep_fees

Enums

DexInstruction
Describes all possible instructions and their required accounts

Functions

cancel_order
Cancel an existing order and remove it from the orderbook.

close_account
Close an inactive and fully settled account

close_market
Close an existing market

consume_events
Crank the processing of DEX events.

create_market
Create a new DEX market

initialize_account
Initialize a new user account

new_order
Execute a new order on the orderbook.

settle
Extract available base and quote token assets from a user account

sweep_fees
Extract accumulated fees from the market. This is an admin instruction